refactored table code
